History object eliminated
