+++ src/main/scala/org/ergoplatform/nodeView/history/storage/modifierprocessors/EmptyBlockSectionProcessor.scala
@@ -1,7 +1,7 @@
 package org.ergoplatform.nodeView.history.storage.modifierprocessors
 
 import org.ergoplatform.modifiers.{BlockSection, ErgoPersistentModifier}
-import scorex.core.consensus.History.ProgressInfo
+import scorex.core.consensus.ProgressInfo
 
 import scala.util.{Failure, Success, Try}
