+++ src/main/scala/scorex/core/consensus/ProgressInfo.scala
@@ -1,0 +1,31 @@
+package scorex.core.consensus
+
+import scorex.core.utils.ScorexEncoder
+import scorex.core.{ModifierTypeId, PersistentNodeViewModifier}
+import scorex.util.ModifierId
+
+/**
+  * Info returned by history to nodeViewHolder after modifier application
+  *
+  * @param branchPoint - branch point in case of rollback
+  * @param toRemove    - modifiers to remove from current node view
+  * @param toApply     - modifiers to apply to current node view
+  * @param toDownload  - modifiers to download from other nodes
+  * @tparam PM - type of used modifier
+  */
+case class ProgressInfo[PM <: PersistentNodeViewModifier](branchPoint: Option[ModifierId],
+                                                          toRemove: Seq[PM],
+                                                          toApply: Seq[PM],
+                                                          toDownload: Seq[(ModifierTypeId, ModifierId)])
+                                                         (implicit encoder: ScorexEncoder) {
+
+  if (toRemove.nonEmpty)
+    require(branchPoint.isDefined, s"Branch point should be defined for non-empty `toRemove`")
+
+  lazy val chainSwitchingNeeded: Boolean = toRemove.nonEmpty
+
+  override def toString: String = {
+    s"ProgressInfo(BranchPoint: ${branchPoint.map(encoder.encodeId)}, " +
+      s" to remove: ${toRemove.map(_.encodedId)}, to apply: ${toApply.map(_.encodedId)})"
+  }
+}
